+++ client_app/src/components/Nodes/NodeClass.svelte.ts
