+++ client_app/src/routes/+page.ts
@@ -1,8 +1,6 @@
 import { analysesMock } from "$lib/analyses";
-import { STORAGE_KEY_ANALYSES, fromLocalStorage } from "$lib/localStorageHandles";
 import { type AnalysesMock } from "$lib/analyses";
 
 export function load(): AnalysesMock {
-    let analysesInLS = fromLocalStorage(STORAGE_KEY_ANALYSES);
-    return analysesInLS ? analysesInLS : analysesMock;
+    return analysesMock;
 }
